PWX-28826: PR review changes and fix portworx_test.go UTs

+++ pkg/controller/storagecluster/storagecluster.go
@@ -423,6 +423,7 @@ func (c *Controller) runPreflightCheck(cluster *corev1.StorageCluster) error {
 			return fmt.Errorf("please make sure your cluster meet all prerequisites and rerun preflight check")
 		}
 		// preflight passed or not required
+		logrus.Infof("preflight check not required")
 		return nil
 	}
 
@@ -439,37 +440,40 @@ func (c *Controller) runPreflightCheck(cluster *corev1.StorageCluster) error {
 	}
 	// TODO: validate cloud permission for other providers as well
 
-	// Create StorageNodes to return pre-flight checks used by c.Driver.Validate()
-	k8sNodeList := &v1.NodeList{}
-	err = c.client.List(context.TODO(), k8sNodeList)
+	// Skip of over driver Validate() if an error has occurred
 	if err == nil {
-		for _, node := range k8sNodeList.Items {
-			logrus.Infof("Create pre-flight storage node entry for node: %s", node.Name)
-			c.createStorageNode(cluster, node.Name)
-		}
-	} else {
-		logrus.WithError(err).Errorf("Failed to get cluster nodes")
-	}
-
-	// Run driver specific pre-flights
-	if err = c.Driver.Validate(toUpdate); err != nil {
-		logrus.WithError(err).Errorf("pre-flight validate failed")
-	}
-
-	// Delete StorageNodes created for c.Driver.Validate() checks.
-	storageNodes := &corev1.StorageNodeList{}
-	err = c.client.List(context.TODO(), storageNodes,
-		&client.ListOptions{Namespace: toUpdate.Namespace})
-	if err == nil {
-		for _, storageNode := range storageNodes.Items {
-			logrus.Infof("Delete validate() storage node entry for node: %s", storageNode.Name)
-			err = c.client.Delete(context.TODO(), storageNode.DeepCopy())
-			if err != nil {
-				logrus.WithError(err).Errorf("failed to delete storage node entry %s: %v", storageNode.Name, err)
+		// Create StorageNodes to return pre-flight checks used by c.Driver.Validate()
+		k8sNodeList := &v1.NodeList{}
+		err = c.client.List(context.TODO(), k8sNodeList)
+		if err == nil {
+			for _, node := range k8sNodeList.Items {
+				logrus.Infof("Create pre-flight storage node entry for node: %s", node.Name)
+				c.createStorageNode(cluster, node.Name)
 			}
+		} else {
+			logrus.WithError(err).Errorf("Failed to get cluster nodes")
 		}
-	} else {
-		logrus.WithError(err).Errorf("Failed to get StorageNodes used for validate.")
+
+		// Run driver specific pre-flights
+		if err = c.Driver.Validate(toUpdate); err != nil {
+			logrus.WithError(err).Errorf("pre-flight validate failed")
+		}
+
+		// Delete StorageNodes created for c.Driver.Validate() checks.
+		storageNodes := &corev1.StorageNodeList{}
+		serr := c.client.List(context.TODO(), storageNodes,
+			&client.ListOptions{Namespace: toUpdate.Namespace})
+		if serr == nil {
+			for _, storageNode := range storageNodes.Items {
+				logrus.Infof("Delete validate() storage node entry for node: %s", storageNode.Name)
+				serr = c.client.Delete(context.TODO(), storageNode.DeepCopy())
+				if serr != nil {
+					logrus.WithError(serr).Errorf("failed to delete storage node entry %s: %v", storageNode.Name, serr)
+				}
+			}
+		} else {
+			logrus.WithError(err).Errorf("Failed to get StorageNodes used for validate.")
+		}
 	}
 
 	condition := &corev1.ClusterCondition{
@@ -477,11 +481,8 @@ func (c *Controller) runPreflightCheck(cluster *corev1.StorageCluster) error {
 		Type:   corev1.ClusterConditionTypePreflight,
 	}
 
-	// XXX - test
-	//err = fmt.Errorf("pre-flighr check force failure")
-
 	if err != nil {
-		logrus.Infof("storage cluster preflight check failed")
+		logrus.Errorf("storage cluster preflight check failed")
 		condition.Status = corev1.ClusterConditionStatusFailed
 		condition.Message = err.Error()
 	} else {
